test: 💍 add some test
